SyntheticLikelihood2

Documentation for SyntheticLikelihood2.

SyntheticLikelihood2.Prior

SyntheticLikelihood2.cov

SyntheticLikelihood2.cov

SyntheticLikelihood2.cut_at

SyntheticLikelihood2.insupport

SyntheticLikelihood2.log_prior_gradient

SyntheticLikelihood2.log_prior_hessian

SyntheticLikelihood2.logpdf

SyntheticLikelihood2.sample_θ

SyntheticLikelihood2.sample_θ

SyntheticLikelihood2.Prior — Type

Struct to hold prior distributions. v is a vector of distributions from Distributions.jl. Univariate and/or multivariate distributions can be used. Each element in the vector is assumed to be independent.

SyntheticLikelihood2.cov — Method

Covariance matrix (1 by 1) for univariate distribution.

SyntheticLikelihood2.cov — Method

Covariance matrix from vector of priors.

SyntheticLikelihood2.cut_at — Method

'Cut' vector at specified indexes, to form a vector of vectors. Note the cut occurs after the index, so [1] would cut between index 1 and 2 in the vector.

SyntheticLikelihood2.insupport — Method

Check if a proposed parameter vector falls within the prior support.

SyntheticLikelihood2.log_prior_gradient — Method

Automatic differentiation to get prior gradient.

SyntheticLikelihood2.log_prior_hessian — Method

Automatic differentiation to get prior Hessian.

SyntheticLikelihood2.logpdf — Method

Evaluate the density of the prior at θ.

SyntheticLikelihood2.sample_θ — Method

Sample a matrix of parameter vectors from the prior

SyntheticLikelihood2.sample_θ — Method

Sample a parameter vector from the prior.