SyntheticLikelihood2

Documentation for SyntheticLikelihood2.

SyntheticLikelihood2.Prior

SyntheticLikelihood2.cut_at

SyntheticLikelihood2.gaussian_simulator

SyntheticLikelihood2.in_prior_support

SyntheticLikelihood2.log_prior_gradient

SyntheticLikelihood2.log_prior_hessian

SyntheticLikelihood2.prior_cov

SyntheticLikelihood2.prior_logpdf

SyntheticLikelihood2.simulate

SyntheticLikelihood2.@loopify

SyntheticLikelihood2.Prior — Type

Struct to hold prior distributions. v is a vector of distributions from Distributions.jl. Univariate and/or multivariate distributions can be used. Each element in the vector is assumed to be independent.

SyntheticLikelihood2.cut_at — Method

'Cut' vector at specified indexes, to form a vector of vectors. Note the cut occurs after the index, so [1] would cut between index 1 and 2 in the vector.

SyntheticLikelihood2.gaussian_simulator — Method

Simulate a three dimensional Gaussian mean vector θ. The covariance is diagonal, and fixed to σ=0.1. Parameter vector θ is the mean vector of the Gaussian

SyntheticLikelihood2.in_prior_support — Method

Check if a parameter vector falls within the prior support.

SyntheticLikelihood2.log_prior_gradient — Method

Automatic differentiation to get prior gradient.

SyntheticLikelihood2.log_prior_hessian — Method

Automatic differentiation to get prior Hessian.

SyntheticLikelihood2.prior_cov — Method

Covariance matrix of the prior.

SyntheticLikelihood2.prior_logpdf — Method

Evaluate the density of the prior at θ.

SyntheticLikelihood2.simulate — Method

Sample parameters from the prior. Note that if n is provided, a matrix is returned with n columns.

SyntheticLikelihood2.@loopify — Macro

Convenience macro to define a method for a simulator that allows simulating in batches. i.e. it takes simulator(rng::AbstractRNG, θ::Vector{Float64}) and wraps it in a for loop to get simulator(rng::AbstractRNG, θ::Matrix{Float64}).